+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fixed cleanup of rojo sourcemap generation process when VSCode exits
 - Fixed color presentations values being unclamped causing errors in other editors
+- Fixed newline not added to separate services and requires when the suggestion imports both at the same time
 
 ## [1.23.0] - 2023-08-06